Refactor out actual Webserver
